modules/accelerators/contextualization/cdf_discovery_aliasing/requirements.txt
cognite-sdk>=7.0
python-dotenv>=1.0
PyYAML>=6.0
pydantic>=2.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
